Refactor for bundling (#38)
