+++ aws-lambda/mds-metrics-sheet/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@aws-lambda/mds-metrics-sheet",
+  "version": "0.1.1",
+  "private": true,
+  "description": "AWS Lambda for MDS Metrics Sheet",
+  "main": "dist/index.js",
+  "scripts": {
+    "build": "tsc --build .",
+    "bundle": "webpack --mode=production --env.npm_package_name=${npm_package_name} --env.npm_package_version=${npm_package_version}",
+    "deploy:us-east-2": "aws lambda update-function-code --region us-east-2 --zip-file fileb://dist/${npm_package_name}.zip --function-name",
+    "deploy:us-west-2": "aws lambda update-function-code --region us-west-2 --zip-file fileb://dist/${npm_package_name}.zip --function-name",
+    "test": "yarn test:prettier && yarn test:eslint && yarn test:unit",
+    "test:eslint": "eslint --ignore-path ../../.gitignore '**/*.ts'",
+    "test:prettier": "prettier --write --check --ignore-path ../../.gitignore '**/*.ts'",
+    "test:unit": "exit 0"
+  },
+  "keywords": [
+    "mds"
+  ],
+  "author": "City of Los Angeles",
+  "dependencies": {
+    "mds-metrics-sheet": "0.0.1"
+  },
+  "devDependencies": {},
+  "license": "Apache-2.0"
+}
