+++ aws-lambda/mds-metrics-sheet/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "../../tsconfig.settings.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "references": [
+    { "path": "../../aws-lambda/aws-serverless-express-handler" },
+    { "path": "../../packages/mds-api-server" },
+    { "path": "../../packages/mds-audit" }
+  ],
+  "include": ["metrics-log.ts", "vehicle-counts.ts"]
+}
