+++ aws-lambda/tsconfig.json
@@ -8,6 +8,7 @@
     { "path": "mds-audit" },
     { "path": "mds-compliance" },
     { "path": "mds-policy" },
-    { "path": "mds-provider" }
+    { "path": "mds-provider" },
+    { "path": "mds-metrics-sheet" }
   ]
 }
