+++ packages/mds-metrics-sheet/metrics-log.ts
@@ -17,9 +17,7 @@
 import GoogleSpreadsheet from 'google-spreadsheet'
 
 import { promisify } from 'util'
-
 import requestPromise from 'request-promise'
-
 import log from 'mds-logger'
 import {
   JUMP_PROVIDER_ID,
@@ -31,8 +29,7 @@ import {
   SHERPA_PROVIDER_ID,
   BOLT_PROVIDER_ID
 } from 'mds-providers'
-
-import { VehicleCountResponse } from './types'
+import { VehicleCountResponse, LastDayStatsResponse, MetricsSheetRow } from './types'
 
 require('dotenv').config()
 
@@ -53,7 +50,16 @@ const creds = {
   private_key: process.env.GOOGLE_PRIVATE_KEY ? process.env.GOOGLE_PRIVATE_KEY.split('\\n').join('\n') : null
 }
 
-async function appendSheet(sheetName: string, rows: ({ date: string; name: string } & unknown)[]) {
+function sum(arr: number[]) {
+  return arr.reduce((total, amount) => total + (amount || 0))
+}
+
+// Round percent to two decimals
+function percent(a: number, total: number) {
+  return Math.round(((total - a) / total) * 10000) / 10000
+}
+
+async function appendSheet(sheetName: string, rows: MetricsSheetRow[]) {
   const doc = new GoogleSpreadsheet(process.env.SPREADSHEET_ID)
   try {
     await promisify(doc.useServiceAccountAuth)(creds)
@@ -72,12 +78,11 @@ async function appendSheet(sheetName: string, rows: ({ date: string; name: strin
   }
 }
 
-async function getProviderMetrics(iter: number): Promise<({ date: string; name: string } & unknown)[]> {
+async function getProviderMetrics(iter: number): Promise<MetricsSheetRow[]> {
   /* after 10 failed iterations, give up */
   if (iter >= 10) {
     throw new Error(`Failed to write to sheet after 10 tries!`)
   }
-
   const token_options = {
     method: 'POST',
     url: `${process.env.AUTH0_DOMAIN}/oauth/token`,
@@ -90,7 +95,6 @@ async function getProviderMetrics(iter: number): Promise<({ date: string; name: 
     },
     json: true
   }
-
   try {
     const token = await requestPromise(token_options)
     const counts_options = {
@@ -98,18 +102,55 @@ async function getProviderMetrics(iter: number): Promise<({ date: string; name: 
       headers: { authorization: `Bearer ${token.access_token}` },
       json: true
     }
+    const last_options = {
+      uri: 'https://api.ladot.io/daily/admin/last_day_stats_by_provider',
+      headers: { authorization: `Bearer ${token.access_token}` },
+      json: true
+    }
 
     const counts: VehicleCountResponse = await requestPromise(counts_options)
-    const rows: ({ date: string; name: string } & unknown)[] = counts
+    const last: LastDayStatsResponse = await requestPromise(last_options)
+
+    const rows: MetricsSheetRow[] = counts
       .filter(p => reportProviders.includes(p.provider_id))
-      .map(row => {
+      .map(provider => {
         const dateOptions = { timeZone: 'America/Los_Angeles', day: '2-digit', month: '2-digit', year: 'numeric' }
         const timeOptions = { timeZone: 'America/Los_Angeles', hour12: false, hour: '2-digit', minute: '2-digit' }
         const d = new Date()
+        let [starts, ends, start_sla, end_sla, telems, telem_sla] = [0, 0, 0, 0, 0, 0]
+        let event_counts = { service_start: 0, provider_drop_off: 0, trip_start: 0, trip_end: 0 }
+        if (last[provider.provider_id].event_counts_last_24h) {
+          event_counts = last[provider.provider_id].event_counts_last_24h
+          starts = last[provider.provider_id].event_counts_last_24h.trip_start || 0
+          ends = last[provider.provider_id].event_counts_last_24h.trip_end || 0
+          telems = last[provider.provider_id].telemetry_counts_last_24h || 0
+          telem_sla = telems ? percent(last[provider.provider_id].late_telemetry_counts_last_24h, telems) : 0
+          start_sla = starts ? percent(last[provider.provider_id].late_event_counts_last_24h.trip_start, starts) : 0
+          end_sla = ends ? percent(last[provider.provider_id].late_event_counts_last_24h.trip_end, ends) : 0
+        }
         return {
           date: `${d.toLocaleDateString('en-US', dateOptions)} ${d.toLocaleTimeString('en-US', timeOptions)}`,
-          name: row.provider,
-          ...row.areas_48h
+          name: provider.provider,
+          registered: provider.count || 0,
+          deployed:
+            sum([
+              provider.status.available,
+              provider.status.unavailable,
+              provider.status.trip,
+              provider.status.reserved
+            ]) || 0,
+          validtrips: 'tbd', // Placeholder for next day valid trip analysis
+          trips: last[provider.provider_id].trips_last_24h || 0,
+          servicestart: event_counts.service_start || 0,
+          providerdropoff: event_counts.provider_drop_off || 0,
+          tripstart: starts,
+          tripend: ends,
+          tripenter: last[provider.provider_id].event_counts_last_24h.trip_enter || 0,
+          tripleave: last[provider.provider_id].event_counts_last_24h.trip_leave || 0,
+          telemetry: telems,
+          telemetrysla: telem_sla,
+          tripstartsla: start_sla,
+          tripendsla: end_sla
         }
       })
     return rows
@@ -119,7 +160,7 @@ async function getProviderMetrics(iter: number): Promise<({ date: string; name: 
   }
 }
 
-exports.handler = () =>
+export const MetricsLogHandler = () =>
   getProviderMetrics(0)
-    .then(rows => appendSheet('Vehicle Counts', rows))
+    .then(rows => appendSheet('Metrics Log', rows))
     .catch((err: Error) => log.error(err))
